requirements.txt
torch==2.5.1
torchvision==0.20.1
faster-coco-eval>=1.6.7
PyYAML
tensorboard
scipy
calflops
transformers
opencv-python
tqdm
scikit-learn
